+++ leaderboard/views.py
@@ -32,11 +32,18 @@ def _get_team_data(request):
     ocelot_team_name = None
     ocelot_team_email = None
     ocelot_team_token = request.session.get('ocelot_team_token')
+    ocelot_team_verified = False
     if ocelot_team_token:
         the_team = Team.objects.get(token=ocelot_team_token)
         ocelot_team_name = the_team.name
         ocelot_team_email = the_team.email
-    return (ocelot_team_name, ocelot_team_email, ocelot_team_token)
+        ocelot_team_verified = the_team.is_verified
+    return (
+        ocelot_team_name,
+        ocelot_team_email,
+        ocelot_team_token,
+        ocelot_team_verified,
+    )
 
 
 def _format_datetime_for_js(stamp):
@@ -118,6 +125,7 @@ def leaderboard(request, competition_id=None):
         ocelot_team_name,
         ocelot_team_email,
         ocelot_team_token,
+        ocelot_team_verified,
     ) = _get_team_data(request)
 
     context = {
@@ -127,6 +135,7 @@ def leaderboard(request, competition_id=None):
         'ocelot_team_name': ocelot_team_name,
         'ocelot_team_email': ocelot_team_email,
         'ocelot_team_token': ocelot_team_token,
+        'ocelot_team_verified': ocelot_team_verified,
     }
     return render(request, 'leaderboard/competition.html', context=context)
 
@@ -163,6 +172,7 @@ def frontpage(request):
         ocelot_team_name,
         ocelot_team_email,
         ocelot_team_token,
+        ocelot_team_verified,
     ) = _get_team_data(request)
 
     context = {
@@ -171,6 +181,7 @@ def frontpage(request):
         'ocelot_team_name': ocelot_team_name,
         'ocelot_team_email': ocelot_team_email,
         'ocelot_team_token': ocelot_team_token,
+        'ocelot_team_verified': ocelot_team_verified,
     }
     return render(request, 'leaderboard/frontpage.html', context=context)
 
@@ -249,10 +260,16 @@ def submit(request):
         ocelot_team_name,
         ocelot_team_email,
         ocelot_team_token,
+        ocelot_team_verified,
     ) = _get_team_data(request)
 
     if not ocelot_team_token:
         _msg = 'You need to be signed in to access this page.'
+        messages.warning(request, _msg)
+        return HttpResponseRedirect('/')
+
+    if not ocelot_team_verified:
+        _msg = 'Your team needs to be verified to access this page.'
         messages.warning(request, _msg)
         return HttpResponseRedirect('/')
 
@@ -278,16 +295,6 @@ def submit(request):
                 return HttpResponseRedirect('/')
             if comp.start_time and current_time <= comp.start_time:
                 _msg = '{0} submission has not started.'.format(comp.name)
-                messages.warning(request, _msg)
-                return HttpResponseRedirect('/')
-
-            if not current_team.is_verified:
-                _msg = (
-                    'The team account {0} has not been verified. '
-                    'Please contact us providing your institution(s) name.'.format(
-                        current_team
-                    )
-                )
                 messages.warning(request, _msg)
                 return HttpResponseRedirect('/')
 
@@ -317,6 +324,11 @@ def submit(request):
                     new_submission.hyp_file.name
                 )
                 messages.success(request, _msg)
+
+                # If is_primary has been selected and the submissions was
+                # successful, update new_submission.is_primary field.
+                if form.cleaned_data['is_primary']:
+                    new_submission.set_primary()  # This implicitly calls save()
             else:
                 _msg = (
                     'Unsuccessful submission of {0}. '
@@ -341,6 +353,7 @@ def submit(request):
         'ocelot_team_name': ocelot_team_name,
         'ocelot_team_email': ocelot_team_email,
         'ocelot_team_token': ocelot_team_token,
+        'ocelot_team_verified': ocelot_team_verified,
     }
     return render(request, 'leaderboard/submission.html', context=context)
 
@@ -352,6 +365,7 @@ def teampage(request):
         ocelot_team_name,
         ocelot_team_email,
         ocelot_team_token,
+        ocelot_team_verified,
     ) = _get_team_data(request)
 
     if not ocelot_team_token:
@@ -435,6 +449,7 @@ def teampage(request):
         'test_set__source_language__code',
         'test_set__target_language__code',
         '-score_chrf',
+        '-date_created',
     )
     for submission in submissions.order_by(*ordering):
         key = submission.test_set
@@ -447,6 +462,16 @@ def teampage(request):
 
         if submission.is_primary:
             primary[key] = submission
+
+    # If no primary system has been selected by the user yet, we will use
+    # the highest-scoring or the latest submission for any given test set.
+    # Based on our ordering defined above, this will be the first object
+    # in the data[key] list, for each of the distinct keys.
+    for key in data.keys():
+        if not key in primary:
+            highest_scoring_or_latest_submission_is_default = data[key][0]
+            highest_scoring_or_latest_submission_is_default.set_primary()
+            primary[key] = highest_scoring_or_latest_submission_is_default
 
     data_triples = []  # (test set, primary submission, all submissions)
     for key in data.keys():
@@ -471,6 +496,7 @@ def teampage(request):
         'ocelot_team_name': ocelot_team_name,
         'ocelot_team_email': ocelot_team_email,
         'ocelot_team_token': ocelot_team_token,
+        'ocelot_team_verified': ocelot_team_verified,
         'publication_name_form': publication_name_form,
         'publication_desc_form': publication_desc_form,
         'publication_survey': publication_survey,
@@ -485,6 +511,7 @@ def updates(request):
         ocelot_team_name,
         ocelot_team_email,
         ocelot_team_token,
+        ocelot_team_verified,
     ) = _get_team_data(request)
 
     context = {
@@ -492,6 +519,7 @@ def updates(request):
         'ocelot_team_name': ocelot_team_name,
         'ocelot_team_email': ocelot_team_email,
         'ocelot_team_token': ocelot_team_token,
+        'ocelot_team_verified': ocelot_team_verified,
     }
     return render(request, 'leaderboard/updates.html', context=context)
 
@@ -503,6 +531,7 @@ def download(request):
         ocelot_team_name,
         ocelot_team_email,
         ocelot_team_token,
+        ocelot_team_verified,
     ) = _get_team_data(request)
 
     context = {
@@ -510,6 +539,7 @@ def download(request):
         'ocelot_team_name': ocelot_team_name,
         'ocelot_team_email': ocelot_team_email,
         'ocelot_team_token': ocelot_team_token,
+        'ocelot_team_verified': ocelot_team_verified,
     }
     return render(request, 'leaderboard/download.html', context=context)
 
@@ -521,6 +551,7 @@ def welcome(request):
         ocelot_team_name,
         ocelot_team_email,
         ocelot_team_token,
+        ocelot_team_verified,
     ) = _get_team_data(request)
 
     if not ocelot_team_token:
@@ -532,5 +563,6 @@ def welcome(request):
         'ocelot_team_name': ocelot_team_name,
         'ocelot_team_email': ocelot_team_email,
         'ocelot_team_token': ocelot_team_token,
+        'ocelot_team_verified': ocelot_team_verified,
     }
     return render(request, 'leaderboard/welcome.html', context=context)